physical-ai-textbook/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0
pydantic-settings>=2.0
sqlmodel>=0.0.8
sqlalchemy>=2.0
asyncpg>=0.27
psycopg2-binary
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
openai>=1.0
qdrant-client[fastembed]>=1.6
better-exceptions
pytest
httpx
alembic>=1.10